+++ sdk/core/azure-core/src/http/curl/curl_session_private.hpp
@@ -198,7 +198,7 @@ namespace Azure { namespace Core { namespace Http {
        * @return Will move the response only if parsing is completed and if the HTTP RawResponse
        * was not moved before.
        */
-      std::unique_ptr<RawResponse> GetResponse()
+      std::unique_ptr<RawResponse> ExtractResponse()
       {
         if (m_parseCompleted && m_response != nullptr)
         {
@@ -404,7 +404,7 @@ namespace Azure { namespace Core { namespace Http {
      * @return the unique ptr to the HTTP RawResponse or null if the HTTP RawResponse is not yet
      * created or was moved before.
      */
-    std::unique_ptr<Azure::Core::Http::RawResponse> GetResponse();
+    std::unique_ptr<Azure::Core::Http::RawResponse> ExtractResponse();
 
     /**
      * @brief Implement #Azure::Core::IO::BodyStream length.