+++ sdk/keyvault/azure-security-keyvault-keys/src/key_client.cpp
@@ -5,8 +5,10 @@
 #include <azure/core/http/http.hpp>
 #include <azure/core/http/policies/policy.hpp>
 
+#include "azure/keyvault/keys/details/key_backup.hpp"
 #include "azure/keyvault/keys/details/key_constants.hpp"
 #include "azure/keyvault/keys/details/key_request_parameters.hpp"
+#include "azure/keyvault/keys/details/key_serializers.hpp"
 #include "azure/keyvault/keys/key_client.hpp"
 
 #include <memory>
@@ -17,6 +19,41 @@ using namespace Azure::Security::KeyVault::Keys;
 using namespace Azure::Core::Http;
 using namespace Azure::Core::Http::Policies;
 using namespace Azure::Core::Http::Policies::_internal;
+
+namespace {
+struct RequestWithContinuationToken
+{
+  std::vector<std::string> Path;
+  std::unique_ptr<std::map<std::string, std::string>> Query;
+};
+
+static inline RequestWithContinuationToken BuildRequestFromContinuationToken(
+    GetSinglePageOptions const& options,
+    std::vector<std::string>&& defaultPath)
+{
+  RequestWithContinuationToken request;
+  request.Path = defaultPath;
+  if (options.ContinuationToken)
+  {
+    // Using a continuation token requires to send the request to the continuation token url instead
+    // of the default url which is used only for the first page.
+    Azure::Core::Url nextPageUrl(options.ContinuationToken.GetValue());
+    request.Query
+        = std::make_unique<std::map<std::string, std::string>>(nextPageUrl.GetQueryParameters());
+    request.Path.clear();
+    request.Path.emplace_back(nextPageUrl.GetPath());
+  }
+  if (options.MaxResults)
+  {
+    if (request.Query == nullptr)
+    {
+      request.Query = std::make_unique<std::map<std::string, std::string>>();
+    }
+    request.Query->emplace("maxResults", std::to_string(options.MaxResults.GetValue()));
+  }
+  return request;
+}
+} // namespace
 
 KeyClient::KeyClient(
     std::string const& vaultUrl,
@@ -50,7 +87,7 @@ Azure::Response<KeyVaultKey> KeyClient::GetKey(
       context,
       Azure::Core::Http::HttpMethod::Get,
       [&name](Azure::Core::Http::RawResponse const& rawResponse) {
-        return _detail::KeyVaultKeyDeserialize(name, rawResponse);
+        return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(name, rawResponse);
       },
       {_detail::KeysPath, name, options.Version});
 }
@@ -66,7 +103,7 @@ Azure::Response<KeyVaultKey> KeyClient::CreateKey(
       Azure::Core::Http::HttpMethod::Post,
       _detail::KeyRequestParameters(keyType, options),
       [&name](Azure::Core::Http::RawResponse const& rawResponse) {
-        return _detail::KeyVaultKeyDeserialize(name, rawResponse);
+        return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(name, rawResponse);
       },
       {_detail::KeysPath, name, "create"});
 }
@@ -81,7 +118,7 @@ Azure::Response<KeyVaultKey> KeyClient::CreateEcKey(
       Azure::Core::Http::HttpMethod::Post,
       _detail::KeyRequestParameters(ecKeyOptions),
       [&keyName](Azure::Core::Http::RawResponse const& rawResponse) {
-        return _detail::KeyVaultKeyDeserialize(keyName, rawResponse);
+        return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(keyName, rawResponse);
       },
       {_detail::KeysPath, keyName, "create"});
 }
@@ -96,7 +133,7 @@ Azure::Response<KeyVaultKey> KeyClient::CreateRsaKey(
       Azure::Core::Http::HttpMethod::Post,
       _detail::KeyRequestParameters(rsaKeyOptions),
       [&keyName](Azure::Core::Http::RawResponse const& rawResponse) {
-        return _detail::KeyVaultKeyDeserialize(keyName, rawResponse);
+        return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(keyName, rawResponse);
       },
       {_detail::KeysPath, keyName, "create"});
 }
@@ -111,9 +148,43 @@ Azure::Response<KeyVaultKey> KeyClient::CreateOctKey(
       Azure::Core::Http::HttpMethod::Post,
       _detail::KeyRequestParameters(octKeyOptions),
       [&keyName](Azure::Core::Http::RawResponse const& rawResponse) {
-        return _detail::KeyVaultKeyDeserialize(keyName, rawResponse);
+        return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(keyName, rawResponse);
       },
       {_detail::KeysPath, keyName, "create"});
+}
+
+Azure::Response<KeyPropertiesSinglePage> KeyClient::GetPropertiesOfKeysSinglePage(
+    GetPropertiesOfKeysSinglePageOptions const& options,
+    Azure::Core::Context const& context) const
+{
+  auto const request = BuildRequestFromContinuationToken(options, {_detail::KeysPath});
+  return m_pipeline->SendRequest<KeyPropertiesSinglePage>(
+      context,
+      Azure::Core::Http::HttpMethod::Get,
+      [](Azure::Core::Http::RawResponse const& rawResponse) {
+        return _detail::KeyPropertiesSinglePageSerializer::KeyPropertiesSinglePageDeserialize(
+            rawResponse);
+      },
+      request.Path,
+      request.Query);
+}
+
+Azure::Response<KeyPropertiesSinglePage> KeyClient::GetPropertiesOfKeyVersionsSinglePage(
+    std::string const& name,
+    GetPropertiesOfKeyVersionsSinglePageOptions const& options,
+    Azure::Core::Context const& context) const
+{
+  auto const request
+      = BuildRequestFromContinuationToken(options, {_detail::KeysPath, name, "versions"});
+  return m_pipeline->SendRequest<KeyPropertiesSinglePage>(
+      context,
+      Azure::Core::Http::HttpMethod::Get,
+      [](Azure::Core::Http::RawResponse const& rawResponse) {
+        return _detail::KeyPropertiesSinglePageSerializer::KeyPropertiesSinglePageDeserialize(
+            rawResponse);
+      },
+      request.Path,
+      request.Query);
 }
 
 Azure::Security::KeyVault::Keys::DeleteKeyOperation KeyClient::StartDeleteKey(
@@ -126,9 +197,24 @@ Azure::Security::KeyVault::Keys::DeleteKeyOperation KeyClient::StartDeleteKey(
           context,
           Azure::Core::Http::HttpMethod::Delete,
           [&name](Azure::Core::Http::RawResponse const& rawResponse) {
-            return _detail::DeletedKeyDeserialize(name, rawResponse);
+            return _detail::DeletedKeySerializer::DeletedKeyDeserialize(name, rawResponse);
           },
           {_detail::KeysPath, name}));
+}
+
+Azure::Security::KeyVault::Keys::RecoverDeletedKeyOperation KeyClient::StartRecoverDeletedKey(
+    std::string const& name,
+    Azure::Core::Context const& context) const
+{
+  return Azure::Security::KeyVault::Keys::RecoverDeletedKeyOperation(
+      m_pipeline,
+      m_pipeline->SendRequest<Azure::Security::KeyVault::Keys::KeyVaultKey>(
+          context,
+          Azure::Core::Http::HttpMethod::Post,
+          [&name](Azure::Core::Http::RawResponse const& rawResponse) {
+            return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(name, rawResponse);
+          },
+          {_detail::DeletedKeysPath, name, "recover"}));
 }
 
 Azure::Response<DeletedKey> KeyClient::GetDeletedKey(
@@ -139,7 +225,118 @@ Azure::Response<DeletedKey> KeyClient::GetDeletedKey(
       context,
       Azure::Core::Http::HttpMethod::Get,
       [&name](Azure::Core::Http::RawResponse const& rawResponse) {
-        return _detail::DeletedKeyDeserialize(name, rawResponse);
+        return _detail::DeletedKeySerializer::DeletedKeyDeserialize(name, rawResponse);
       },
       {_detail::DeletedKeysPath, name});
+}
+
+Azure::Response<DeletedKeySinglePage> KeyClient::GetDeletedKeysSinglePage(
+    GetDeletedKeysSinglePageOptions const& options,
+    Azure::Core::Context const& context) const
+{
+  auto const request = BuildRequestFromContinuationToken(options, {_detail::DeletedKeysPath});
+  return m_pipeline->SendRequest<DeletedKeySinglePage>(
+      context,
+      Azure::Core::Http::HttpMethod::Get,
+      [](Azure::Core::Http::RawResponse const& rawResponse) {
+        return _detail::KeyPropertiesSinglePageSerializer::DeletedKeySinglePageDeserialize(
+            rawResponse);
+      },
+      request.Path,
+      request.Query);
+}
+
+Azure::Response<PurgedKey> KeyClient::PurgeDeletedKey(
+    std::string const& name,
+    Azure::Core::Context const& context) const
+{
+  return m_pipeline->SendRequest<PurgedKey>(
+      context,
+      Azure::Core::Http::HttpMethod::Delete,
+      [](Azure::Core::Http::RawResponse const&) { return PurgedKey(); },
+      {_detail::DeletedKeysPath, name});
+}
+
+Azure::Response<KeyVaultKey> KeyClient::UpdateKeyProperties(
+    KeyProperties const& properties,
+    Azure::Nullable<std::list<KeyOperation>> const& keyOperations,
+    Azure::Core::Context const& context) const
+{
+  return m_pipeline->SendRequest<KeyVaultKey>(
+      context,
+      Azure::Core::Http::HttpMethod::Patch,
+      _detail::KeyRequestParameters(properties, keyOperations),
+      [&properties](Azure::Core::Http::RawResponse const& rawResponse) {
+        return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(properties.Name, rawResponse);
+      },
+      {_detail::KeysPath, properties.Name, properties.Version});
+}
+
+Azure::Response<std::vector<uint8_t>> KeyClient::BackupKey(
+    std::string const& name,
+    Azure::Core::Context const& context) const
+{
+  // Use the internal model KeyBackup to parse from Json
+  auto response = m_pipeline->SendRequest<_detail::KeyBackup>(
+      context,
+      Azure::Core::Http::HttpMethod::Post,
+      [](Azure::Core::Http::RawResponse const& rawResponse) {
+        return _detail::KeyBackup::Deserialize(rawResponse);
+      },
+      {_detail::KeysPath, name, "backup"});
+
+  // Convert the internal KeyBackup model to a raw vector<uint8_t>.
+  return Azure::Response<std::vector<uint8_t>>(
+      response.ExtractValue().Value, response.ExtractRawResponse());
+}
+
+Azure::Response<KeyVaultKey> KeyClient::RestoreKeyBackup(
+    std::vector<uint8_t> const& backup,
+    Azure::Core::Context const& context) const
+{
+  _detail::KeyBackup backupModel;
+  backupModel.Value = backup;
+  return m_pipeline->SendRequest<KeyVaultKey>(
+      context,
+      Azure::Core::Http::HttpMethod::Post,
+      backupModel,
+      [](Azure::Core::Http::RawResponse const& rawResponse) {
+        return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(rawResponse);
+      },
+      {_detail::KeysPath, "restore"});
+}
+
+Azure::Response<KeyVaultKey> KeyClient::ImportKey(
+    std::string const& name,
+    JsonWebKey const& keyMaterial,
+    Azure::Core::Context const& context) const
+{
+  ImportKeyOptions const importKeyOptions(name, keyMaterial);
+  return m_pipeline->SendRequest<KeyVaultKey>(
+      context,
+      Azure::Core::Http::HttpMethod::Put,
+      [&importKeyOptions]() {
+        return _detail::ImportKeyOptionsSerializer::ImportKeyOptionsSerialize(importKeyOptions);
+      },
+      [&name](Azure::Core::Http::RawResponse const& rawResponse) {
+        return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(name, rawResponse);
+      },
+      {_detail::KeysPath, name});
+}
+
+Azure::Response<KeyVaultKey> KeyClient::ImportKey(
+    ImportKeyOptions const& importKeyOptions,
+    Azure::Core::Context const& context) const
+{
+  return m_pipeline->SendRequest<KeyVaultKey>(
+      context,
+      Azure::Core::Http::HttpMethod::Put,
+      [&importKeyOptions]() {
+        return _detail::ImportKeyOptionsSerializer::ImportKeyOptionsSerialize(importKeyOptions);
+      },
+      [&importKeyOptions](Azure::Core::Http::RawResponse const& rawResponse) {
+        return _detail::KeyVaultKeySerializer::KeyVaultKeyDeserialize(
+            importKeyOptions.Name(), rawResponse);
+      },
+      {_detail::KeysPath, importKeyOptions.Name()});
 }